refactor(services/importer): use new utility function to associate with exercise
